+++ src/main/java/lotto/exception/InvalidMoneyAmountException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.exception;
 
 public class InvalidMoneyAmountException extends RuntimeException {
     public InvalidMoneyAmountException(String message) {
