+++ src/test/java/lotto/domain/lottery/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottery;
 
 import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -11,6 +11,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import lotto.exception.InvalidLottoNumberException;
 
 class LottoTicketTest {
     @DisplayName("로또 번호를 입력하여 로또 한 장 생성")
