+++ src/test/java/lotto/domain/purchase/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.purchase;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import lotto.exception.InvalidMoneyAmountException;
 
 class MoneyTest {
     @DisplayName("로또 구입 금액을 입력하여 인스턴스를 생성한다.")
